Create Counting_occurrence_of_each_letter_in_list_of_random-characters
Counts the occurrences of each letter in the list. To do so, it creates a list named 
counts that has 26 int values, each of which counts the occurrences of a letter. That is, 
counts[0] counts the number of times 'a' appears in the list, counts[1] counts the number 
of time 'b' appears, and so on.